refactor: package as paperfinder_nabi
