+++ README.md
@@ -1,0 +1,22 @@
+# paperfinder_nabi
+
+`paperfinder_nabi` collects recent papers from several RSS feeds and uses
+Google's Gemini model to recommend a short list that matches your research
+interests.
+
+## Installation
+
+```bash
+pip install paperfinder_nabi
+```
+
+## Usage
+
+After installing, run the module with your Gemini API key:
+
+```bash
+python -m paperfinder_nabi --api_key YOUR_API_KEY
+```
+
+The package ships with default RSS feeds, keywords and previously liked paper
+examples located in `paperfinder_nabi/data/`.
